multi-wallet (#5337)
* Update send export

* Update nav controller

* Remove Destination

* Update address input

* Rename

* Code clean

* Search text field

* Update

* Address input

* Update add address flow

* Update destination menu

* Update scan by qrcode

* Hide keyboard

* Delete menu

* Update address handle

* Update to web3 account

* Fix layout scroll

* Update token select

* Fix scanned text change

* Value input

* Support fee

* Support memo

* Support fee select

* Receive selection

* Support web3 to wallet

* Update menu style

* Fix nav controller

* Refactor: Implement TransferContactBottomSheetDialogFragment for user selection

* Update strings

* Update

* Remove useless code

* Fix layout

* Fix fee check

* Fix web3 input

* Hide address menu

* Fix transfer pop back stack

* Fix delete address return

* Update UI after fee switching

* Fix number format

* Update sub title

* Fix fee id

* Update layout

* Remove address check

* Update sub title

* User transfer

* Support scheme

* Update hide deposit

* Fix typo

* Update web3 symbol

* Return empty id

* Check web3 gas

* More sync utxo

* Refactor: Use consistent circle style for PIN input

* Prepare check

* Update address withdrawal

* Check Not Confirmed UTXO

* Wallet menu

* Update wallet card

* Mock wallet database

* Update pin view

* Wallet init

* Replace chain icon

* Fix navigation

* Update classic

* Update transaction details

* Refresh job

* Update transaction page

* Update sol stake

* Fix layout

* Subtract gas

* Transaction UI

* Wallet card bind data

* Update

* Update color bar

* Save select

* More menu

* Navigation to web3 transaction

* Web3 search

* Check local address

* Update schema

* Update

* Update web3 transaction

* Web3 filter

* Update web3 token hidden

* Update type filter

* Update scheme

* Update

* Update wallet token

* Update

* Bump com.reown:android-bom from 1.1.2 to 1.3.0

Bumps [com.reown:android-bom](https://github.com/reown-con/reown-kotlin) from 1.1.2 to 1.3.0.
- [Commits](https://github.com/reown-con/reown-kotlin/commits)

---
updated-dependencies:
- dependency-name: com.reown:android-bom
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* Web3 receive

* Rename and code clean

* Update menu

* Fetch wallets

* Fix web3 send

* Update swap subtitle

* Search and market

* Update

* Update

* Safely convert balance string using toBigDecimalOrNull

* Fix address sync

* Fix wallet id

* Update

* Update distribution

* Text overflow

* Update UI

* Fix sql

* Update holder

* Fix error

* Fix lost token pirce

* Update

* Recursive fetch

* Update asset distribution

* Update

* Classic receive

* Update refresh transaction

* Update

* Update

* Update scheme

* Update scheme

* Update

* Replace field name

* Filter sol

* Fix swap

* Feature/rpc (#5390)

* update build gradle

* Bump com.appsflyer:af-android-sdk from 6.16.1 to 6.16.2 (#5373)

Bumps [com.appsflyer:af-android-sdk](https://bitbucket.org/appsflyer/appsflyer-android-sdk) from 6.16.1 to 6.16.2.
- [Commits](https://bitbucket.org/appsflyer/appsflyer-android-sdk/commits)

---
updated-dependencies:
- dependency-name: com.appsflyer:af-android-sdk
  dependency-type: direct:production
  update-type: version-update:semver-patch
...


* update splash screen duration

* Estimate erc-20 fee

* Update Sol priority fee

* Replace tip gas

---------


* Update

* tweak use gas

* Update layout

* Update header margin

* Handle error

* Update distribution layout

* Remove useless code

* Remove useless code

* Fix balance change

* Fix swap quote

* Replace balance change token

* Fix balance click

* Use route rpc

* Remove web3 api

* Revert test code

* Refactor Web3TokenListBottomSheetDialogFragment with remote search support

* Address validate

* Support send external url

* Update api

* Update sub title

* Hide progress

* Update compose

* Rename

* Code clean

* Input visual transformation

* Recent web3 token

* Find token from mxin

* Update scheme

* Update

* Check chain

* Update web3 search

* Update layout

* Update wallet pin check

* fix omit repalce rpc usage

* Fix sql

* Revert distrbution

* Web3 raw transcation

* Update

* Update tx state

* Update web3 transcation

* update post tx

* Update json

* Update

* Update type

* Refresh transaction

* update estimate fee request

* Update

* Scan external url

* Fix  chain id

* Check pending transaction

* Update

* Revert "Bump kotlinVersion from 2.1.10 to 2.1.20 (#5401)"

This reverts commit 58f55fe87878f7aa66c08352b9a9d7d98fff3b53.

* fix proxy rpc return 'null'

* Hide id

* use simulate api instead of parse

* Update pending layout

* Fix layout

* Update dark icon

* Update

* Update balance to zero for missing assets

* Update button when on error

* Update

* Update

* Update

* Fix type menu

* Update schema

---------

+++ app/src/main/java/one/mixin/android/api/response/web3/QuoteResult.kt
@@ -18,8 +18,8 @@ fun QuoteResult?.rate(fromToken: SwapToken?, toToken: SwapToken?): BigDecimal {
     if (this == null) return BigDecimal.ZERO
     if (fromToken == null || toToken == null) return BigDecimal.ZERO
     return runCatching {
-        val inValue = fromToken.realAmount(inAmount)
-        val outValue = toToken.realAmount(outAmount)
+        val inValue = fromToken.realAmount(inAmount.toBigDecimal())
+        val outValue = toToken.realAmount(outAmount.toBigDecimal())
         outValue.divide(inValue, 8, RoundingMode.CEILING)
     }.getOrDefault(BigDecimal.ZERO)
 }
